ref(ui): Move file into correct view folder
Since this component is not used outside the releases view and is not
designed to be a highly reusable component, it should be under views not
components.
